fix(data init): externalize initDataSet() method
Move initDataSet() methodfrom `DataInitializer`  to a separate bean
`InitService` and call `InitService.initDataSet()` to enable
transactional behavior  (the transactional method and the
ApplicationReadyEvent listener can't be in the same class).

Add the `@Profile` annotation on `DataInitializer` to create this bean
only when the "hibernate-init" profile is active.

Add rollbackFor attribute on `initDataSet()` to ensure rollback is
triggered for any Exception in method `initDataSet()`.