feat(role): Add RoleNotFoundException
Add class RoleNotFoundException to handle cases where a role is not
found in the database.

In ApiMessage, add constant `ROLE_NOT_FOUND` to represent the error
message.

In UserService, propagate the exception in `registerNewUserAccount()`.

In UserController, propagate the exception in `register()`.

In GlobalControllerExceptionHandler, add method
`handleRoleNotFoundException()` to return a 404 Not Found response with
an ApiError DTO.